[inkling] Flatten the ops package into a single ops.py
Merge ops/kernel/{attention,precision_aligned_ops,rel_score_mod,triton_sconv,
triton_swiglu}.py and ops/cp_utils.py into one ops.py (kimi_k3 layout), and
move the multimodal helpers (vision_encoder, image_processing,
audio_processing) up next to their consumers. Model-specific public ops gain
the inkling_ prefix: inkling_swiglu_fp32, inkling_sum_fp32, inkling_sconv_fp32,
inkling_sconv_fp32_packed, inkling_sp_residual_conv, inkling_te_attention.
