+++ miles_plugins/models/inkling/image_processing.py
