--- miles_plugins/models/inkling/ops/kernel/rel_score_mod.py
@@ -1,48 +1,0 @@
-from collections.abc import Callable
-from functools import cache
-
-try:
-    import cutlass.cute as cute
-    from cutlass.cute import Float32
-    from flash_attn.cute.seqlen_info import SeqlenInfoQK
-except Exception as _import_error:
-    cute = None
-    Float32 = None
-    SeqlenInfoQK = None
-    _cute_import_error = _import_error
-else:
-    _cute_import_error = None
-
-
-@cache
-def get_inkling_relative_attention_score_mod(rel_extent: int) -> Callable:
-    if cute is None or Float32 is None or SeqlenInfoQK is None:
-        raise ImportError(
-            "Inkling relative attention requires the vendored FA4 CUTE interface."
-        ) from _cute_import_error
-
-    @cute.jit
-    def score_mod_rel_bias(
-        scores: cute.TensorSSA,
-        b_idx: cute.TensorSSA,
-        h_idx: cute.TensorSSA,
-        q_idx: cute.TensorSSA,
-        kv_idx: cute.TensorSSA,
-        seqlen_info: SeqlenInfoQK,
-        aux_tensors: list[cute.Tensor],
-    ) -> cute.TensorSSA:
-        rel_logits = aux_tensors[0]
-
-        seqlen_local_offset = seqlen_info.seqlen_k - seqlen_info.seqlen_q
-        rel_dist = (q_idx + seqlen_local_offset) - kv_idx
-        global_q_idx = seqlen_info.offset_q + q_idx
-
-        rel_dist_0 = rel_dist[0]
-        rel_idx = rel_dist_0 if rel_dist_0 >= 0 else 0
-        rel_idx = rel_idx if rel_idx < rel_extent else (rel_extent - 1)
-
-        rel_bias = rel_logits[global_q_idx[0], h_idx[0], rel_idx]
-        rel_bias = Float32(rel_bias) if rel_dist_0 == rel_idx else Float32(0.0)
-        return scores + rel_bias
-
-    return score_mod_rel_bias
